Group DOM predicates in separate file
