+++ packages/alfa-dom/tsconfig.json
@@ -6,6 +6,7 @@
     "src/index.ts",
     "src/jsx.ts",
     "src/namespace.ts",
+    "src/common/predicate.ts",
     "src/node.ts",
     "src/node/attribute.ts",
     "src/node/comment.ts",
